Interpolate a sequence of values with Easing or Stepping Given Data Points

This function generates a sequence of values based on a specified easing or stepping function. It supports linear, polynomial, exponential, and other smooth transitions, as well as stepped transitions.

Usage

seq_data(data, type = "linear", step_count = NULL, ease = NULL)

Arguments

data

Numeric vector, matrix, data frame, or list. The input data to be used for generating the sequence.

type

Character string specifying the type of sequence. Supported types include:

"linear": Linear interpolation.

"quad": Quadratic easing.

"cubic": Cubic easing.

"quart": Quartic easing.

"quint": Quintic easing.

"exp": Exponential easing.

"circle": Circular easing.

"back": Back easing with overshoot.

"elastic": Elastic easing with oscillation.

"sine": Sine wave easing.

"bounce": Bouncing easing.

"step": Stepped transitions.

Defaults to "linear".

step_count

Integer specifying the number of steps for the "step" type. Must be between 1 and the length of data. Defaults to NULL.

ease

Character string specifying the direction of easing. Supported values are:

"in": Easing starts slow and accelerates.

"out": Easing starts fast and decelerates.

"in_out": Easing combines both behaviors.

Applicable only for non-linear types. Defaults to NULL.

Value

A numeric vector containing the generated sequence.

For "linear", a smoothly interpolated sequence is returned.

For "step", a sequence with distinct steps is generated.

For other easing types, the sequence follows the specified smooth transition curve.

Details

The seq_data function calculates a sequence of values based on the specified type and ease. The data input is used to determine the range (minimum and maximum) of the sequence to then be interpolated, and the resulting sequence is normalized between 0 and 1 before applying the specified easing or stepping function.

For "step" type, the number of steps can be controlled using step_count. The ease parameter has no effect when type is "linear" or "step".

Note

This function supports various easing functions commonly used in animations and graphics, as well as stepped sequences for discrete transitions. Invalid or unsupported inputs will result in informative error messages or warnings.

Examples

# Generate a linear sequence
seq_data(1:10, type = "linear")
#>  [1]  1  2  3  4  5  6  7  8  9 10

# Generate a quadratic easing sequence
seq_data(rnorm(100,14,5), type = "quad", ease = "in_out")
#>   [1]  0.9383283  0.9438051  0.9602353  0.9876190  1.0259562  1.0752468
#>   [7]  1.1354909  1.2066885  1.2888396  1.3819442  1.4860022  1.6010138
#>  [13]  1.7269788  1.8638972  2.0117692  2.1705946  2.3403735  2.5211059
#>  [19]  2.7127918  2.9154311  3.1290239  3.3535702  3.5890700  3.8355233
#>  [25]  4.0929300  4.3612902  4.6406039  4.9308711  5.2320917  5.5442658
#>  [31]  5.8673934  6.2014745  6.5465091  6.9024971  7.2694386  7.6473336
#>  [37]  8.0361821  8.4359840  8.8467395  9.2684484  9.7011108 10.1447266
#>  [43] 10.5992960 11.0648188 11.5412951 12.0287248 12.5271081 13.0364448
#>  [49] 13.5567350 14.0879787 14.6274375 15.1586812 15.6789714 16.1883081
#>  [55] 16.6866914 17.1741211 17.6505974 18.1161203 18.5706896 19.0143055
#>  [61] 19.4469678 19.8686767 20.2794322 20.6792341 21.0680826 21.4459776
#>  [67] 21.8129191 22.1689071 22.5139417 22.8480228 23.1711504 23.4833245
#>  [73] 23.7845451 24.0748123 24.3541260 24.6224862 24.8798929 25.1263462
#>  [79] 25.3618460 25.5863923 25.7999851 26.0026244 26.1943103 26.3750427
#>  [85] 26.5448216 26.7036470 26.8515190 26.9884375 27.1144025 27.2294140
#>  [91] 27.3334720 27.4265766 27.5087277 27.5799253 27.6401694 27.6894601
#>  [97] 27.7277972 27.7551809 27.7716111 27.7770879

# Generate a stepped sequence with 5 steps
seq_data(rpois(100,3), type = "step", step_count = 5)
#>   [1] 0.0 0.0 0.0 0.0 0.0 0.0 0.0 0.0 0.0 0.0 1.4 1.4 1.4 1.4 1.4 1.4 1.4 1.4
#>  [19] 1.4 1.4 1.4 1.4 1.4 1.4 1.4 1.4 1.4 1.4 1.4 1.4 2.8 2.8 2.8 2.8 2.8 2.8
#>  [37] 2.8 2.8 2.8 2.8 2.8 2.8 2.8 2.8 2.8 2.8 2.8 2.8 2.8 2.8 4.2 4.2 4.2 4.2
#>  [55] 4.2 4.2 4.2 4.2 4.2 4.2 4.2 4.2 4.2 4.2 4.2 4.2 4.2 4.2 4.2 4.2 5.6 5.6
#>  [73] 5.6 5.6 5.6 5.6 5.6 5.6 5.6 5.6 5.6 5.6 5.6 5.6 5.6 5.6 5.6 5.6 5.6 5.6
#>  [91] 7.0 7.0 7.0 7.0 7.0 7.0 7.0 7.0 7.0 7.0